TSUGI-11 Instructor and student view finished

+++ actions/Delete.php
@@ -1,7 +1,6 @@
 <?php
 require_once "../../config.php";
 require_once "../dao/QW_DAO.php";
-require_once "../util/Utils.php";
 
 use \Tsugi\Core\LTIX;
 use \QW\DAO\QW_DAO;
@@ -13,20 +12,26 @@ $p = $CFG->dbprefix;
 
 $QW_DAO = new QW_DAO($PDOX, $p);
 
-$SetID=$_SESSION["SetID"];
-$QID=$_GET["QID"];
+$SetID = $_SESSION["SetID"];
+$QID = isset($_GET["QID"]) ? $_GET["QID"] : -1;
 
 if ( $USER->instructor ) {
 
-    $QW_DAO->deleteQuestion($QID);
+    if ($QID > -1) {
+        // Delete all answers
+        $QW_DAO->deleteAnswersToQuestion($QID);
 
-    $remainingQ = $QW_DAO->getQuestions($SetID);
-	$QNum = 0;
-    foreach ( $remainingQ as $question ) {
-        $QNum++;
-        $QW_DAO->updateQNumber($question["QID"], $QNum);
+        // Delete the quesiton
+        $QW_DAO->deleteQuestion($QID);
+
+        // Fix numbering
+        $remainingQ = $QW_DAO->getQuestions($SetID);
+        $QNum = 0;
+        foreach ( $remainingQ as $question ) {
+            $QNum++;
+            $QW_DAO->updateQNumber($question["QID"], $QNum);
+        }
     }
 
-	
-    header( 'Location: '.addSession('../instructor-home.php?Add=0') ) ;
+    header( 'Location: '.addSession('../instructor-home.php') ) ;
 }
